Fix: project structure and deleting extra complexity
