+++ src/main/java/com/mateo/ShoppingCart/Marketplace/model/ShoppingCartOutput.java
@@ -1,0 +1,15 @@
+package com.mateo.ShoppingCart.Marketplace.model;
+
+import com.mateo.ShoppingCart.Marketplace.domain.ShoppingCart;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+public class ShoppingCartOutput {
+    private ShoppingCart shoppingCart;
+}
