+++ src/main/java/com/mateo/ShoppingCart/Marketplace/services/ShoppingCartServices.java
@@ -1,18 +1,45 @@
 package com.mateo.ShoppingCart.Marketplace.services;
 
+/* La capa de servicio es aquella dedicada a manejar la logica del negocio
+ *  de tal modo que podamos gestionar las transacciones necesarias para el
+ *  repository*/
 
 import com.mateo.ShoppingCart.Marketplace.domain.*;
 import com.mateo.ShoppingCart.Marketplace.repository.ShoppingCartRepository;
+import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
 import java.time.Instant;
+import java.util.List;
 
+@Service
 public class ShoppingCartServices {
+    private ShoppingCartRepository repository;
 
-    private ShoppingCartRepository shoppingCartRepository;
+    public ShoppingCartServices(ShoppingCartRepository repository){
+        this.repository = repository;
+    }
 
-    public ShoppingCartServices(ShoppingCartRepository shoppingCartRepository){
-        this.shoppingCartRepository = shoppingCartRepository;
+    public List<Product> getAllProducts(){
+        return repository.getAllProducts();
+    }
+
+    public Product getProductById(ProductId id){
+        return repository.getProductById(id);
+    }
+
+    public Product createProduct(Product product){
+        repository.createProduct(product);
+        return product;
+    }
+
+    public Product updateProductById(ProductId id, Product product){
+        repository.updateProductById(id, product);
+
+        return product;
+    }
+
+    public void deleteProductById(ProductId id){
+        repository.deleteProductById(id);
     }
 
     public ShoppingCart addProductToShoppingCart(Product product, ShoppingCart shoppingCart){
@@ -25,16 +52,16 @@ public class ShoppingCartServices {
 
         //Creamos la instancia actualizada
         ShoppingCart updatedShoppingCart = new ShoppingCart(shoppingCart.getClientId(),
-                                                            shoppingCart.getCreatedAt(),
-                                                            shoppingCart.getUpdatedAt(),
-                                                            shoppingCart.getProducts()
-                                                            );
+                shoppingCart.getCreatedAt(),
+                shoppingCart.getUpdatedAt(),
+                shoppingCart.getProducts()
+        );
 
         //Calculamos el nuevo valor total
         updatedShoppingCart.setTotal(shoppingCart.calculateTotalPrice(updatedShoppingCart.getProducts()));
 
         //Modificamos el producto en el repositorio
-        shoppingCartRepository.addProductToShoppingCart(product.getProductId(), shoppingCart.getClientId());
+        repository.addProductToShoppingCart(product.getProductId(), shoppingCart.getClientId());
 
         return updatedShoppingCart;
     }
@@ -46,9 +73,9 @@ public class ShoppingCartServices {
 
         //Creamos la instancia actualizada a retornar
         ShoppingCart updatedShoppingCart = new ShoppingCart(shoppingCart.getClientId(),
-                                                            shoppingCart.getCreatedAt(),
-                                                            shoppingCart.getCreatedAt(),
-                                                            shoppingCart.getProducts());
+                shoppingCart.getCreatedAt(),
+                shoppingCart.getCreatedAt(),
+                shoppingCart.getProducts());
 
         //Actualizamos el valor total del carrito
         updatedShoppingCart.setTotal(shoppingCart.calculateTotalPrice(updatedShoppingCart.getProducts()));
@@ -56,7 +83,7 @@ public class ShoppingCartServices {
         //Actualizamos la fecha de modificacion
         updatedShoppingCart.setUpdatedAt(Instant.now());
 
-        shoppingCartRepository.removeProductFromShoppingCart(id);
+        repository.removeProductFromShoppingCart(id);
 
         return updatedShoppingCart;
     }
@@ -64,9 +91,9 @@ public class ShoppingCartServices {
     public ShoppingCart increaseProductQuantity(ProductId id, ShoppingCart shoppingCart){
 
         ShoppingCart updatedShoppingCart = new ShoppingCart(shoppingCart.getClientId(),
-                                                            shoppingCart.getCreatedAt(),
-                                                            shoppingCart.getUpdatedAt(),
-                                                            shoppingCart.getProducts());
+                shoppingCart.getCreatedAt(),
+                shoppingCart.getUpdatedAt(),
+                shoppingCart.getProducts());
 
         Product foundProduct = updatedShoppingCart.getProducts().get(id);
 
@@ -78,7 +105,7 @@ public class ShoppingCartServices {
 
         //Creamos un nuevo objeto money que sera usado en la operacion del repository
         Money updatedPrice = new Money(foundProduct.getPrice().getValue(), Badge.USD);
-        shoppingCartRepository.updateProductQuantity(foundProduct.getProductId(), updatedPrice, foundProduct.getQuantity());
+        repository.updateProductQuantity(foundProduct.getProductId(), updatedPrice, foundProduct.getQuantity());
 
         //Actualizamos el valor del total del shopping cart
         updatedShoppingCart.getProducts().remove(id);
@@ -93,9 +120,9 @@ public class ShoppingCartServices {
 
     public ShoppingCart decreaseProductQuantity(ProductId id, ShoppingCart shoppingCart){
         ShoppingCart updatedShoppingCart = new ShoppingCart(shoppingCart.getClientId(),
-                                                            shoppingCart.getCreatedAt(),
-                                                            shoppingCart.getUpdatedAt(),
-                                                            shoppingCart.getProducts());
+                shoppingCart.getCreatedAt(),
+                shoppingCart.getUpdatedAt(),
+                shoppingCart.getProducts());
 
         Product foundProduct = updatedShoppingCart.getProducts().get(id);
         //Validamos si el item a eliminar tiene solo 1 unidad
@@ -110,7 +137,7 @@ public class ShoppingCartServices {
             foundProduct.setPriceByQuantity(foundProduct.getQuantity());
             Money updatedPrice = new Money(foundProduct.getPrice().getValue(), Badge.USD);
 
-            shoppingCartRepository.updateProductQuantity(foundProduct.getProductId(), updatedPrice, foundProduct.getQuantity());
+            repository.updateProductQuantity(foundProduct.getProductId(), updatedPrice, foundProduct.getQuantity());
 
             //Actualizamos el producto dentro del shopping cart
             updatedShoppingCart.getProducts().remove(id);
